client to send multicast s-message

+++ api/oc_knx.c
@@ -974,7 +974,7 @@ oc_is_s_mode_request(oc_request_t *request)
   // TODO: add code to check if the request is part of the same setup
   // e.g. check idd of the device vs the data in the request
 
-  PRINT("  oc_is_s_mode_request %.*s\n", request->uri_path_len,
+  PRINT("  oc_is_s_mode_request %.*s\n", (int)request->uri_path_len,
         request->uri_path);
   if (strncmp(".knx", request->uri_path, request->uri_path_len) == 0) {
     return true;
